Refactor namespaces to match folder structure.
