+++ WundergroundNetConsole/Program.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WundergroundNetLib;
-
+using WundergroundNetLib.Model;
 using WundergroundNetLib.Interfaces;
 
 namespace WundergroundNetConsole
